+++ index.php
@@ -2,10 +2,14 @@
 require_once('vendor/autoload.php');
 use Oc\Blog\controller\UserController;
 use Oc\Blog\model\UserModel;
-
+use Oc\Blog\controller\TwigService;
 // On veut afficher les utilisateurs
+
 $controller = new UserController();
 $controller->showUsers();
+
+$twigService = new TwigService(); 
+$controller = new UserController($twigService);
 
 //$controller->displayText();
 
@@ -17,4 +21,3 @@ $controller->showUsers();
 //foreach($users as $user){
 //    echo "$user[pseudo]";
 //}
-
