+++ src/controller/UserController.php
@@ -11,22 +11,10 @@ use Twig\Loader\FilesystemLoader;
 
 class UserController
 {
-    private Environment $twig;
+    private TwigService $twig;
 
-    public function _construct()
+    public function _construct(TwigService $twig)
     {
-        // On créé le systeme de fichier Twig pour retrouver les vues (html)
-        $loader = new FilesystemLoader('src/view');
-
-        // On configure twig (on ajoute le mode "debug" et on supprime le "cache")
-        $twig = new Environment($loader, [
-            'debug' => true,
-            'cache' => false //__DIR__ .'/tmp'
-        ]);
-
-        // On active le var_dump() de twig pour debugger
-        $twig->addExtension(new DebugExtension());
-
         // Je stock la configuration twig dans notre variable twig du controller
         $this->twig = $twig;
     }
@@ -34,7 +22,7 @@ class UserController
 
     public function displayText(){
         printf("hello de UserController() \n ");
-        $this->twig->render('home.twig');
+        $this->twig->getTwig()->render('home.twig');
     }
 
     /**
@@ -43,16 +31,16 @@ class UserController
      */
     public function showUsers()
     {
-//        $userModel = new UserModel();
-//        $users = $userModel->getUsers();
-        $users = [
+        $userModel = new UserModel();
+        $users = $userModel->getUsers();
+        /*$users = [
             [
                 'id' => 1,
                 'pseudo' => 'Djazaanti'
             ],
-        ];
-
-        $this->twig->render('home.twig', [$users]);
+        ];*/
+        $this->twig->getTwig()->render('home.html.twig', [$users]);
+        var_dump($this->twig);
     }
 
 }
